endless version

+++ Project2/SequenceAssignHandler.java
@@ -2,14 +2,14 @@ import org.apache.thrift.TException;
 import sequenceassign.*;
 
 public class SequenceAssignHandler implements SequenceAssign.Iface{
-	private long currentSequence;
+	private int currentSequence;
 
 	public SequenceAssignHandler(){
 		currentSequence = 1;
 	}
 
 	@Override
-	public synchronized long sequenceAssign() {
+	public synchronized int sequenceAssign() {
         return currentSequence++;
 	}
