+++ Project2/sequenceassign.thrift
@@ -1,5 +1,5 @@
 namespace java sequenceassign
 
 service SequenceAssign {
-    i64 sequenceAssign(),
+    i32 sequenceAssign(),
 }
